requirements.txt
watchdog==3.0.0
schedule==1.2.0
requests==2.31.0
GitPython==3.1.40
pre-commit==2.20.0
black==22.3.0
flake8==4.0.1
isort==5.10.1